requirements.txt
tensorflow==1.12.0
tensorflow-hub==0.1.1
runway-python==0.0.22
scipy==1.2.0